Rename "funcall" to "funcalls".
