+++ ailets-rs/gpt/src/funcalls.rs
